+++ plugin-packs/postcss-preset-env/src/index.js
@@ -1,174 +1,31 @@
 import autoprefixer from 'autoprefixer';
-import browserslist from 'browserslist';
 import cssdb from 'cssdb';
-import { pluginsById as plugins } from './lib/plugins-by-id';
-import getTransformedInsertions from './lib/get-transformed-insertions';
-import getUnsupportedBrowsersByFeature from './lib/get-unsupported-browsers-by-feature';
-import idsByExecutionOrder from './lib/ids-by-execution-order';
-import writeToExports from './lib/write-to-exports';
-import getOptionsForBrowsersByFeature from './lib/get-options-for-browsers-by-feature';
-import { pluginIdHelp } from './lib/plugin-id-help';
-import { pluginHasSideEffects } from './lib/plugins-with-side-effects';
-import { log, dumpLogs, resetLogger } from './lib/log-helper';
-import { featuresWithClientSide } from './lib/plugins-with-client-side';
-import logFeaturesList from './lib/log-features-list';
+import writeToExports from './side-effects/write-to-exports.mjs';
+import { pluginIdHelp } from './plugins/plugin-id-help.mjs';
+import { dumpLogs, resetLogger } from './log/helper.mjs';
+import logFeaturesList from './log/features-list.mjs';
+import { listFeatures } from './lib/list-features.mjs';
+import { initializeSharedOptions } from './lib/shared-options.mjs';
 
-const DEFAULT_STAGE = 2;
-const OUT_OF_RANGE_STAGE = 5;
-
-const plugin = opts => {
+const plugin = (opts) => {
 	// initialize options
 	const options = Object(opts);
-	const features = Object(options.features);
-	const enableClientSidePolyfills = 'enableClientSidePolyfills' in options ? options.enableClientSidePolyfills : true;
-	const featureNamesInOptions = Object.keys(features);
-	const insertBefore = Object(options.insertBefore);
-	const insertAfter = Object(options.insertAfter);
+	const featureNamesInOptions = Object.keys(Object(options.features));
 	const browsers = options.browsers;
-	let stage = DEFAULT_STAGE;
+	const sharedOptions = initializeSharedOptions(options);
 
-	resetLogger();
-
-	if (typeof options.stage !== 'undefined') {
-		if (options.stage === false) {
-			stage = OUT_OF_RANGE_STAGE;
-		} else {
-			stage = Math.min(parseInt(options.stage, 10), OUT_OF_RANGE_STAGE) || 0;
-		}
-	}
-
-	if (stage === OUT_OF_RANGE_STAGE) {
-		log('Stage has been disabled, features will be handled via the "features" option.');
-	} else {
-		log(`Using features from Stage ${stage}`);
-	}
-
-	const autoprefixerOptions = options.autoprefixer;
-	const sharedOpts = initializeSharedOpts(options);
-	const stagedAutoprefixer = autoprefixerOptions === false
-		? () => {}
-		: autoprefixer(Object.assign({ overrideBrowserslist: browsers }, autoprefixerOptions));
-
-	// polyfillable features (those with an available postcss plugin)
-	const polyfillableFeatures = cssdb.concat(
-		// additional features to be inserted before cssdb features
-		getTransformedInsertions(insertBefore, 'insertBefore'),
-		// additional features to be inserted after cssdb features
-		getTransformedInsertions(insertAfter, 'insertAfter'),
-	).filter(
-		// inserted features or features with an available postcss plugin
-		feature => feature.insertBefore || feature.id in plugins,
-	).sort(
-		// features sorted by execution order and then insertion order
-		(a, b) => idsByExecutionOrder.indexOf(a.id) - idsByExecutionOrder.indexOf(b.id) || (a.insertBefore ? -1 : b.insertBefore ? 1 : 0) || (a.insertAfter ? 1 : b.insertAfter ? -1 : 0),
-	).map(
-		// polyfillable features as an object
-		feature => {
-			// target browsers for the polyfill
-			const unsupportedBrowsers = getUnsupportedBrowsersByFeature(feature);
-
-			return feature.insertBefore || feature.insertAfter ? {
-				browsers: unsupportedBrowsers,
-				plugin:   feature.plugin,
-				id:       `${feature.insertBefore ? 'before' : 'after'}-${feature.id}`,
-				stage:    OUT_OF_RANGE_STAGE + 1, // So they always match
-			} : {
-				browsers: unsupportedBrowsers,
-				plugin:   plugins[feature.id],
-				id:       feature.id,
-				stage:    feature.stage,
-			};
-		},
-	);
-
-	// staged features (those at or above the selected stage)
-	const stagedFeatures = polyfillableFeatures.filter(feature => {
-		const isAllowedStage = feature.stage >= stage;
-		const isAllowedByType = enableClientSidePolyfills || !featuresWithClientSide.includes(feature.id);
-		const isDisabled = features[feature.id] === false;
-		const isAllowedFeature = features[feature.id] ? features[feature.id] : isAllowedStage && isAllowedByType;
-
-		if (isDisabled) {
-			log(`  ${feature.id} has been disabled by options`);
-		} else if (!isAllowedStage) {
-			if (isAllowedFeature) {
-				log(`  ${feature.id} has been enabled by options`);
-			} else {
-				log(`  ${feature.id} with stage ${feature.stage} has been disabled`);
-			}
-		} else if (!isAllowedByType) {
-			log(`  ${feature.id} has been disabled by "enableClientSidePolyfills: false".`);
-		}
-
-		return isAllowedFeature;
-	}).map(
-		feature => {
-			let options;
-			let plugin;
-
-			options = getOptionsForBrowsersByFeature(browsers, feature, cssdb);
-
-			if (features[feature.id] === true) {
-				// if the plugin is enabled
-				options = sharedOpts ? Object.assign({}, options, sharedOpts) : undefined;
-			} else {
-				options = sharedOpts
-					// if the plugin has shared options and individual options
-					? Object.assign({}, options, sharedOpts, features[feature.id])
-					// if the plugin has individual options
-					: Object.assign({}, options, features[feature.id]);
-			}
-
-			if (feature.plugin.postcss) {
-				plugin = feature.plugin(options);
-			} else {
-				plugin = feature.plugin;
-			}
-
-			return {
-				browsers: feature.browsers,
-				plugin: plugin,
-				pluginOptions: options,
-				id: feature.id,
-			};
-		},
-	);
-
-	// browsers supported by the configuration
-	const supportedBrowsers = browserslist(browsers, { ignoreUnknownVersions: true });
-
-	// - features supported by the stage
-	// - features with `true` or with options
-	// - required for the browsers
-	// - having "importFrom" or "exportTo" options
-	const supportedFeatures = stagedFeatures.filter((feature) => {
-		if (feature.id in features) {
-			return features[feature.id];
-		}
-
-		if (pluginHasSideEffects(feature)) {
-			return true;
-		}
-
-		const unsupportedBrowsers = browserslist(feature.browsers, {
-			ignoreUnknownVersions: true,
-		});
-
-		const needsPolyfill = supportedBrowsers.some(supportedBrowser => {
-			return unsupportedBrowsers.some(unsupportedBrowser => unsupportedBrowser === supportedBrowser);
-		});
-
-		if (!needsPolyfill) {
-			log(`${feature.id} disabled due to browser support`);
-		}
-
-		return needsPolyfill;
+	const features = listFeatures(cssdb, options, sharedOptions);
+	const plugins = features.map((feature) => {
+		return feature.plugin;
 	});
 
-	const usedPlugins = supportedFeatures.map(feature => feature.plugin);
-	usedPlugins.push(stagedAutoprefixer);
+	if (options.autoprefixer !== false) {
+		plugins.push(
+			autoprefixer(Object.assign({ overrideBrowserslist: browsers }, options.autoprefixer)),
+		);
+	}
 
-	logFeaturesList(supportedFeatures, options);
+	logFeaturesList(features, options);
 
 	const internalPlugin = () => {
 		return {
@@ -180,8 +37,11 @@ const plugin = opts => {
 					dumpLogs(result);
 				}
 
+				// Always reset the logger, if when debug is false
+				resetLogger();
+
 				if (options.exportTo) {
-					writeToExports(sharedOpts.exportTo, opts.exportTo);
+					writeToExports(sharedOptions.exportTo, opts.exportTo);
 				}
 			},
 		};
@@ -191,34 +51,8 @@ const plugin = opts => {
 
 	return {
 		postcssPlugin: 'postcss-preset-env',
-		plugins: [...usedPlugins, internalPlugin()],
+		plugins: [...plugins, internalPlugin()],
 	};
-};
-
-const initializeSharedOpts = opts => {
-	if ('importFrom' in opts || 'exportTo' in opts || 'preserve' in opts) {
-		const sharedOpts = {};
-
-		if ('importFrom' in opts) {
-			sharedOpts.importFrom = opts.importFrom;
-		}
-
-		if ('exportTo' in opts) {
-			sharedOpts.exportTo = {
-				customMedia: {},
-				customProperties: {},
-				customSelectors: {},
-			};
-		}
-
-		if ('preserve' in opts) {
-			sharedOpts.preserve = opts.preserve;
-		}
-
-		return sharedOpts;
-	}
-
-	return false;
 };
 
 plugin.postcss = true;