+++ plugin-packs/postcss-preset-env/src/log/helper.mjs
@@ -15,3 +15,4 @@ export function dumpLogs(result) {
 
 	resetLogger();
 }
+