+++ plugin-packs/postcss-preset-env/src/test/test.mjs
@@ -1,0 +1,9 @@
+// import './options-matrix.mjs';
+import './lib/feature-is-inserted-or-has-plugin.mjs';
+import './lib/feature-is-less.mjs';
+import './lib/format-staged-feature.mjs';
+import './lib/get-unsupported-browsers-by-feature.mjs';
+import './lib/list-features/list-features.mjs';
+import './lib/prepare-features-list.mjs';
+import './lib/stage.mjs';
+import './lib/transformed-insertions.mjs';